文章を復帰させるようにした、あとリファクタ
