+++ front/src/components/Form/Box.tsx
@@ -6,20 +6,22 @@ import { TextPosition } from '../../types/LayoutProps';
 export interface BoxProps {
   name: string;
   type: TextPosition;
+  body?: string;
 }
 
-const InputText: FC<BoxProps> = ({ name }) => (
+const InputText: FC<BoxProps> = ({ name, body }) => (
   <Form.Field
     required
     control={Input}
     name={name}
     data-cy={name}
     label="本文"
+    value={body || undefined}
   />
 );
 
-const Box: FC<BoxProps> = ({ ...args }) => {
-  const { name, type } = args;
+const Box: FC<BoxProps> = ({ ...props }) => {
+  const { name, type } = props;
 
   return (
     <>
@@ -30,27 +32,27 @@ const Box: FC<BoxProps> = ({ ...args }) => {
           data-position={type}
           id={`column-${name}`}
         >
-          <InputText {...args} />
+          <InputText {...props} />
         </Grid.Column>
       ) : (
         <Grid columns={2} data-position={type} id={`column-${name}`}>
           {type === TextPosition.LEFT ? (
             <>
               <Grid.Column centered="true">
-                <InputText {...args} />
+                <InputText {...props} />
               </Grid.Column>
               <Grid.Column centered="true">
-                <InputImage {...args} />
+                <InputImage {...props} />
               </Grid.Column>
             </>
           ) : null}
           {type === TextPosition.RIGHT ? (
             <>
               <Grid.Column centered="true">
-                <InputImage {...args} />
+                <InputImage {...props} />
               </Grid.Column>
               <Grid.Column centered="true">
-                <InputText {...args} />
+                <InputText {...props} />
               </Grid.Column>
             </>
           ) : null}
