+++ front/src/components/Form/SectionTable.tsx
@@ -1,92 +1,7 @@
-import React, {
-  FC,
-  useState,
-  SyntheticEvent,
-  useReducer,
-  Reducer,
-} from 'react';
-import {
-  Grid,
-  GridRow,
-  GridColumn,
-  Menu,
-  Button,
-  Icon,
-  Segment,
-} from 'semantic-ui-react';
-import Box from './Box';
-import {
-  TextPosition,
-  SectionListProps,
-  Section,
-} from '../../types/LayoutProps';
-
-interface RowProps {
-  activeItem: string;
-  name: string;
-}
-
-const InputlayoutRow = ({ activeItem, name }: RowProps) => (
-  <Grid.Row>
-    {activeItem === 'text_only' ? (
-      <Box name={name} type={TextPosition.CENTER} />
-    ) : null}
-    {activeItem === 'left_text' ? (
-      <Box name={name} type={TextPosition.LEFT} />
-    ) : null}
-    {activeItem === 'right_text' ? (
-      <Box name={name} type={TextPosition.RIGHT} />
-    ) : null}
-  </Grid.Row>
-);
-
-const InputLayoutChangeMenu: FC<{}> = ({ children }) => {
-  return (
-    <GridRow>
-      <GridColumn>
-        <Menu>
-          <Menu.Item header>レイアウトを選択</Menu.Item>
-          {children}
-        </Menu>
-      </GridColumn>
-    </GridRow>
-  );
-};
-
-interface SectionBarProps {
-  name: string;
-}
-const SectionBar: FC<SectionBarProps> = ({ name }) => {
-  const [activeItem, setActiveItem] = useState('text_only');
-
-  const handleMenuChange = (e: SyntheticEvent, type: string) => {
-    e.preventDefault();
-    setActiveItem(type);
-  };
-
-  const menuItems = [
-    { body: '文章のみ', type: 'text_only' },
-    { body: '文章と画像', type: 'left_text' },
-    { body: '画像と文章', type: 'right_text' },
-  ].map((value, index) => (
-    <Menu.Item
-      key={index.toString()}
-      name={value.type}
-      active={activeItem === value.type}
-      data-cy={value.type}
-      onClick={(e: SyntheticEvent) => handleMenuChange(e, value.type)}
-    >
-      {value.body}
-    </Menu.Item>
-  ));
-
-  return (
-    <Grid divided>
-      <InputLayoutChangeMenu>{menuItems}</InputLayoutChangeMenu>
-      <InputlayoutRow activeItem={activeItem} name={name} />
-    </Grid>
-  );
-};
+import React, { FC, SyntheticEvent, useReducer, Reducer } from 'react';
+import { Grid, Button, Icon, Segment } from 'semantic-ui-react';
+import { SectionListProps, Section } from '../../types/LayoutProps';
+import SectionBar from './SectionBar';
 
 // ここのstateは1つ前のイベントのオブジェクト
 const reducer: Reducer<StateType, ActionType> = (state, action) => {
@@ -120,6 +35,7 @@ type ActionType =
 interface Item {
   id: number;
   name: string;
+  body?: string;
 }
 interface StateType {
   items: Item[];
@@ -133,6 +49,7 @@ const SectionTable: FC<SectionListProps> = ({ sections }) => {
       const i = initSections.map((value, index) => ({
         id: index + 1,
         name: `section${index + 1}`,
+        body: value.body,
       }));
 
       initialState = {
@@ -153,9 +70,9 @@ const SectionTable: FC<SectionListProps> = ({ sections }) => {
   return (
     <>
       <Grid celled="internally" columns={2}>
-        {state.items.map((item: { id: number }) => (
-          <Segment key={item.id.toString()} data-cy={`inputSection${item.id}`}>
-            <SectionBar name={`section${item.id}`} />
+        {state.items.map(item => (
+          <Segment key={item.id} data-cy={`inputSection${item.id}`}>
+            <SectionBar name={`section${item.id}`} body={item.body} />
 
             {/* TODO: 消すときにホワンとさせたいかも */}
             <Button
